feat: coin sources are actually logged and saved to file and trhe functionaly exists

+++ src/main/kotlin/xyz/nibblz/islandeconomist/Save.kt
@@ -8,7 +8,7 @@ import java.nio.file.Files
 @Serializable
 data class CoinChange(
     var amount: Int = 0,
-    var timestamp: Int = 0,
+    var timestamp: Long = 0,
     var source: String = "Unknown"
 )
